requirements.txt
flask>=3.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
gunicorn>21.2.0
